+++ now.json
@@ -1,0 +1,7 @@
+{
+  "version": 2,
+  "builds": [
+    { "src": "package.json", "use": "@now/next" },
+    { "src": "*.js", "use": "@now/node" }
+  ]
+}
