+++ package.json
@@ -4,9 +4,10 @@
   "description": "> **ATENÇÃO:** Antes de começar, se [cadastre na vaga](https://linxneemuchaordic.recruiterbox.com/) via Recruiterbox **=]**",
   "main": "index.js",
   "scripts": {
-    "dev": "node server.js",
+    "dev": "next",
     "build": "next build",
-    "start": "NODE_ENV=production node server.js",
+    "start": "NODE_ENV=production next start",
+    "deploy": "now --prod",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -21,17 +22,16 @@
   },
   "homepage": "https://github.com/PauloStanize/frontend-developer-challenge#readme",
   "dependencies": {
-    "@zeit/next-sass": "^1.0.1",
-    "axios": "^0.19.0",
-    "dotenv": "^8.1.0",
-    "express": "^4.17.1",
-    "formik": "^1.5.8",
-    "next": "^9.0.7",
-    "node-sass": "^4.12.0",
-    "react": "^16.10.1",
-    "react-dom": "^16.10.1",
-    "reactstrap": "^8.0.1",
-    "sass-loader": "^8.0.0",
-    "styled-jsx": "^3.2.2"
+    "@zeit/next-sass": "1.0.1",
+    "axios": "0.19.0",
+    "dotenv": "8.1.0",
+    "express": "4.17.1",
+    "formik": "1.5.8",
+    "next": "9.0.7",
+    "node-sass": "4.12.0",
+    "react": "16.10.1",
+    "react-dom": "16.10.1",
+    "sass-loader": "8.0.0",
+    "styled-jsx": "3.2.2"
   }
 }
