--- server.js
@@ -1,26 +1,0 @@
-const express = require('express')
-const next = require('next')
-
-const port = parseInt(process.env.PORT, 10) || 3000
-const dev = process.env.NODE_ENV !== 'production'
-const app = next({ dev })
-const handle = app.getRequestHandler()
-
-const envConfig = require('./envConfig')
-
-app.prepare().then(() => {
-  const server = express()
-
-  server.get("/:page", (req, res) => {
-    return app.render(req, res, "/", { page: req.params.page })
-  })
-
-  server.all('*', (req, res) => {
-    return handle(req, res)
-  })
-
-  server.listen(port, err => {
-    if (err) throw err
-    console.log(`> Ready on http://localhost:${envConfig['port']}`)
-  })
-})
